+++ src/rp2_common/hardware_spi/spi.c
@@ -73,7 +73,7 @@ uint spi_get_baudrate(const spi_inst_t *spi) {
 
 // Write len bytes from src to SPI. Simultaneously read len bytes from SPI to dst.
 // Note this function is guaranteed to exit in a known amount of time (bits sent * time per bit)
-size_t __not_in_flash_func(spi_write_read_blocking)(spi_inst_t *spi, const uint8_t *src, uint8_t *dst, size_t len) {
+int __not_in_flash_func(spi_write_read_blocking)(spi_inst_t *spi, const uint8_t *src, uint8_t *dst, size_t len) {
     invalid_params_if(SPI, 0 > (int)len);
 
     // Never have more transfers in flight than will fit into the RX FIFO,
@@ -92,11 +92,11 @@ size_t __not_in_flash_func(spi_write_read_blocking)(spi_inst_t *spi, const uint8
         }
     }
 
-    return len;
+    return (int)len;
 }
 
 // Write len bytes directly from src to the SPI, and discard any data received back
-size_t __not_in_flash_func(spi_write_blocking)(spi_inst_t *spi, const uint8_t *src, size_t len) {
+int __not_in_flash_func(spi_write_blocking)(spi_inst_t *spi, const uint8_t *src, size_t len) {
     invalid_params_if(SPI, 0 > (int)len);
     // Write to TX FIFO whilst ignoring RX, then clean up afterward. When RX
     // is full, PL022 inhibits RX pushes, and sets a sticky flag on
@@ -118,14 +118,14 @@ size_t __not_in_flash_func(spi_write_blocking)(spi_inst_t *spi, const uint8_t *s
     // Don't leave overrun flag set
     spi_get_hw(spi)->icr = SPI_SSPICR_RORIC_BITS;
 
-    return len;
+    return (int)len;
 }
 
 // Read len bytes directly from the SPI to dst.
 // repeated_tx_data is output repeatedly on SO as data is read in from SI.
 // Generally this can be 0, but some devices require a specific value here,
 // e.g. SD cards expect 0xff
-size_t __not_in_flash_func(spi_read_blocking)(spi_inst_t *spi, uint8_t repeated_tx_data, uint8_t *dst, size_t len) {
+int __not_in_flash_func(spi_read_blocking)(spi_inst_t *spi, uint8_t repeated_tx_data, uint8_t *dst, size_t len) {
     invalid_params_if(SPI, 0 > (int)len);
     const size_t fifo_depth = 8;
     size_t rx_remaining = len, tx_remaining = len;
@@ -141,11 +141,11 @@ size_t __not_in_flash_func(spi_read_blocking)(spi_inst_t *spi, uint8_t repeated_
         }
     }
 
-    return len;
+    return (int)len;
 }
 
 // Write len halfwords from src to SPI. Simultaneously read len halfwords from SPI to dst.
-size_t __not_in_flash_func(spi_write16_read16_blocking)(spi_inst_t *spi, const uint16_t *src, uint16_t *dst, size_t len) {
+int __not_in_flash_func(spi_write16_read16_blocking)(spi_inst_t *spi, const uint16_t *src, uint16_t *dst, size_t len) {
     invalid_params_if(SPI, 0 > (int)len);
     // Never have more transfers in flight than will fit into the RX FIFO,
     // else FIFO will overflow if this code is heavily interrupted.
@@ -163,11 +163,11 @@ size_t __not_in_flash_func(spi_write16_read16_blocking)(spi_inst_t *spi, const u
         }
     }
 
-    return len;
+    return (int)len;
 }
 
 // Write len bytes directly from src to the SPI, and discard any data received back
-size_t __not_in_flash_func(spi_write16_blocking)(spi_inst_t *spi, const uint16_t *src, size_t len) {
+int __not_in_flash_func(spi_write16_blocking)(spi_inst_t *spi, const uint16_t *src, size_t len) {
     invalid_params_if(SPI, 0 > (int)len);
     // Deliberately overflow FIFO, then clean up afterward, to minimise amount
     // of APB polling required per halfword
@@ -187,12 +187,12 @@ size_t __not_in_flash_func(spi_write16_blocking)(spi_inst_t *spi, const uint16_t
     // Don't leave overrun flag set
     spi_get_hw(spi)->icr = SPI_SSPICR_RORIC_BITS;
 
-    return len;
+    return (int)len;
 }
 
 // Read len halfwords directly from the SPI to dst.
 // repeated_tx_data is output repeatedly on SO as data is read in from SI.
-size_t __not_in_flash_func(spi_read16_blocking)(spi_inst_t *spi, uint16_t repeated_tx_data, uint16_t *dst, size_t len) {
+int __not_in_flash_func(spi_read16_blocking)(spi_inst_t *spi, uint16_t repeated_tx_data, uint16_t *dst, size_t len) {
     invalid_params_if(SPI, 0 > (int)len);
     const size_t fifo_depth = 8;
     size_t rx_remaining = len, tx_remaining = len;
@@ -208,5 +208,5 @@ size_t __not_in_flash_func(spi_read16_blocking)(spi_inst_t *spi, uint16_t repeat
         }
     }
 
-    return len;
+    return (int)len;
 }
